pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-source-collector"
version = "0.1.0"
description = "Fetch balances across Tinkoff Investments, Bybit, Binance, and Interactive Brokers with a unified interface."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27,<1.0",
    "pydantic>=2.6,<3.0",
    "python-dotenv>=1.0,<2.0",
    "typer>=0.12,<1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<8.0",
    "pytest-cov>=4.1,<5.0",
    "ruff>=0.4,<0.5",
    "mypy>=1.10,<2.0"
]

[project.scripts]
portfolio-balances = "portfolio_source_collector.cli.main:app"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

